+++ tf_tutorial_2_keras/tf-tutorial-2-keras.py
@@ -1,0 +1,24 @@
+import os
+import collections
+import sys
+import keras
+import math
+import numpy as np
+import tensorflow as tf
+
+from keras.models import Sequential
+from keras.layers import Dense,Dropout,Activation,Flatten,Input
+from keras.layers import Embedding,LSTM
+from keras.utils import np_utils
+from keras.utils import to_categorical
+from keras.optimizers import SGD
+from keras.models import Model
+from keras.layers import Input, Dense,InputLayer
+
+
+
+
+
+
+
+
